Fix a bug in TFXlaCallModuleOpToStablehloPass regarding PlatformIndexArg handling
The issue is that the "PlatformIndexArg" of a StableHLO module is not always a noop argument as was originally expected. When a StableHLO module contains function calls inside, this "PlatformIndexArg" will be propagated along the call graph. Therefore, unconditional removing this arg will remove a still-being-used SSA value and trigger an assertion.

The fix is that instead of removing the arg on the callee function side, we can add a dummy I32 operand on the caller side. After the inlining, this dummy operand will be dead code eliminated and produce the same result as before.

PiperOrigin-RevId: 705166253